fix: prometheus disable multiproc
