+++ bentoml/testing/pytest/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from .plugin import TEST_MODEL_CONTEXT
+
+__all__ = ["TEST_MODEL_CONTEXT"]
